requirements.txt
# Dependências para Instagram Investigator com Tkinter
requests>=2.28.0
python-dotenv>=1.0.0
pyinstaller>=5.13.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Dependências para processamento de texto e análise
nltk>=3.7.0
scikit-learn>=1.0.0
numpy>=1.22.0
pillow>=9.0.0

# Dependências para integração com outras redes sociais
tweepy>=4.14.0
facebook-sdk>=3.1.0
linkedin-api>=0.1.5
tiktokapi>=0.1.0
pandas>=2.0.0
networkx>=3.1.0

# Dependências para exportação
reportlab>=4.0.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0